+++ esp32/src/encoder/NewEncoderAdapter.hpp
@@ -6,7 +6,7 @@
 
 class NewEncoderAdapter : public Encoder{
     private:
-        const uint16_t PPR;
+        const uint8_t PPR;
         const uint8_t CLK_PIN;
         const uint8_t DT_PIN;
         const int8_t MAX_POS;
@@ -16,7 +16,7 @@ class NewEncoderAdapter : public Encoder{
         uint16_t counterClockwiseRevolutions;
         static void rotatedEncoder(NewEncoder* newEncoder, const volatile NewEncoder::EncoderState* state, void* newEncoderAdapter);
     public:
-        NewEncoderAdapter(uint8_t clkPin, uint8_t dtPin, uint16_t ppr);
+        NewEncoderAdapter(uint8_t clkPin, uint8_t dtPin, uint8_t ppr);
         int8_t getPosition() override;
         uint16_t getClockwiseRevolutions() override;
         uint16_t getCounterclockwiseRevolutions() override;
